feat: 블럭 환불가능기간 7일 점검 로직 생성

+++ src/main/java/com/blockpage/blockservice/application/service/PaymentService.java
@@ -67,7 +67,7 @@ public class PaymentService implements PaymentUseCase {
             case KAKAO_PAY_REFUND -> {
                 PaymentEntityDto paymentEntityDto = paymentPersistencePort.getPaymentByOrderId(query.getOrderId());
                 Block orderedBlock = blockPersistencePort.getBlockByOrderId(query.getOrderId());
-                if (orderedBlock.getBlockQuantity().equals(paymentEntityDto.getBlockQuantity())) {
+                if (orderedBlock.getBlockQuantity().equals(paymentEntityDto.getBlockQuantity()) && orderedBlock.isAvailableRefund()) {
                     KakaoPayRefundParams kakaoPayRefundParams = KakaoPayRefundParams.addEssentialParams(paymentEntityDto);
                     KakaoPayRefundDto kakaoPayRefundDto = paymentRequestPort.cancel(kakaoPayRefundParams);
                     blockPersistencePort.deleteBlockByOrderId(kakaoPayRefundDto.getOrderId());
